Added Integration with the log file from time-traq-nag
Another file can now be dropped in the app. The distinction is done via the file name.
The data is passed as prop to the Timeline component.
Added the nags as a first group (even before the ALL) but hidden by defaul. When the nag log is added, then the group is populated.
Nag group population happens in a separate useEffect hook. There, the nag file lines are processed and the extracted nag items are added to the group. Finally, the group is made visible.