Fixed a bug with deserialization.
The TaskInfo objects were just plain objects and so no functions of it worked.
Added the logic to create the objects if they had type:TaskInfo.
Added a migration script to migrate affected storage files.
The script should be placed inside the .storage folder of the corresponding application. We also need json5-wildcard-patch.js from the project of the same name (bundle with `yarn bundle`)