Fixes for bugs and vulnerabilities
Fixed a bug when the second line of the time-traq contains "Personal" and generally with calling the useEffect one more time (seemingly for not reason). The reason was that I modified the dataProp (as part of dataToAppend and dataToExtract) so I'm now making a copy inside the useEffect.
Fixed a bug when reading the settings file doesn't allow adding more tasks: parsing the TaskInfo class was not available and the task didn't have a type.
For the parsing, I had to expose the parseRecursive function and to add some base array/object cases.
Fixed some critical vulnerabilities by upgrading the yarn packages. babel/core was the main issue, so i removed the core folder and the lines for core from yarn.lock and ran yarn, but that adjusted more packages. Well, it works afterwards and the critical vulnerabilities are gone, so it's fine.